@@ -13,58 +13,58 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.1.3",
-    "@angular/cdk": "^13.1.3",
-    "@angular/common": "~13.1.3",
-    "@angular/compiler": "~13.1.3",
-    "@angular/core": "~13.1.3",
-    "@angular/forms": "~13.1.3",
-    "@angular/material": "^13.1.3",
-    "@angular/platform-browser": "~13.1.3",
-    "@angular/platform-browser-dynamic": "~13.1.3",
-    "@angular/router": "~13.1.3",
-    "@angular/service-worker": "~13.1.3",
+    "@angular/animations": "^14.2.0",
+    "@angular/cdk": "^14.2.0",
+    "@angular/common": "^14.2.0",
+    "@angular/compiler": "^14.2.0",
+    "@angular/core": "^14.2.0",
+    "@angular/forms": "^14.2.0",
+    "@angular/material": "^14.2.0",
+    "@angular/platform-browser": "^14.2.0",
+    "@angular/platform-browser-dynamic": "^14.2.0",
+    "@angular/router": "^14.2.0",
+    "@angular/service-worker": "^14.2.0",
     "@rxweb/types": "^1.0.8",
-    "angular-cli-ghpages": "^1.0.0",
+    "angular-cli-ghpages": "^1.0.3",
     "d3": "^6.7.0",
-    "dexie": "^2.0.4",
+    "dexie": "^3.2.2",
     "ngx-filesize": "^2.0.16",
     "rxjs": "^6.6.7",
-    "tslib": "^2.3.1",
-    "zone.js": "~0.11.4"
+    "tslib": "^2.4.0",
+    "zone.js": "^0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.2.5",
-    "@angular-eslint/builder": "^12.7.0",
-    "@angular-eslint/eslint-plugin": "^12.7.0",
-    "@angular-eslint/eslint-plugin-template": "^12.7.0",
-    "@angular-eslint/schematics": "^12.7.0",
-    "@angular-eslint/template-parser": "^12.7.0",
-    "@angular/cli": "^13.1.4",
-    "@angular/compiler-cli": "~13.1.3",
-    "@angular/language-service": "~13.1.3",
+    "@angular-devkit/build-angular": "^14.2.1",
+    "@angular-eslint/builder": "^14.0.3",
+    "@angular-eslint/eslint-plugin": "^14.0.3",
+    "@angular-eslint/eslint-plugin-template": "^14.0.3",
+    "@angular-eslint/schematics": "^14.0.3",
+    "@angular-eslint/template-parser": "^14.0.3",
+    "@angular/cli": "^14.2.1",
+    "@angular/compiler-cli": "^14.2.0",
+    "@angular/language-service": "^14.2.0",
     "@types/d3": "^6.7.5",
     "@types/file-saver": "^2.0.5",
-    "@types/jasmine": "^3.10.3",
+    "@types/jasmine": "^3.10.6",
     "@types/jasminewd2": "~2.0.10",
-    "@types/node": "^12.20.42",
-    "@typescript-eslint/eslint-plugin": "^5.10.0",
-    "@typescript-eslint/parser": "^5.10.0",
+    "@types/node": "^12.20.55",
+    "@typescript-eslint/eslint-plugin": "^5.29.0",
+    "@typescript-eslint/parser": "^5.29.0",
     "codelyzer": "^6.0.2",
-    "eslint": "^7.32.0",
-    "eslint-plugin-import": "2.25.4",
+    "eslint": "^8.18.0",
+    "eslint-plugin-import": "2.25.2",
     "eslint-plugin-jsdoc": "30.7.13",
     "eslint-plugin-prefer-arrow": "1.2.3",
-    "jasmine-core": "~3.99.0",
+    "jasmine-core": "^3.99.1",
     "jasmine-spec-reporter": "~5.0.2",
-    "karma": "^6.3.17",
-    "karma-chrome-launcher": "~3.1.0",
+    "karma": "^6.4.0",
+    "karma-chrome-launcher": "^3.1.1",
     "karma-coverage-istanbul-reporter": "~3.0.3",
-    "karma-jasmine": "~4.0.1",
+    "karma-jasmine": "^4.0.2",
     "karma-jasmine-html-reporter": "^1.7.0",
     "protractor": "~7.0.0",
-    "ts-node": "^10.4.0",
-    "typescript": "^4.5.5"
+    "ts-node": "^10.9.1",
+    "typescript": "^4.8.2"
   },
   "prettier": {
     "trailingComma": "es5",
